enfin

+++ work/filev6.c
@@ -21,15 +21,13 @@ int filev6_open(const struct unix_filesystem *u, uint16_t inr, struct filev6 *fv
  * @return >0: the number of bytes of the file read; 0: end of file;
  *             the appropriate error code (<0) on error
  */
+//TODO check cast (uint32_t)
 int filev6_readblock(struct filev6 *fv6, void *buf){
     M_REQUIRE_NON_NULL(fv6);
     M_REQUIRE_NON_NULL(buf);
 
     int remaining = inode_getsize(&fv6->i_node) - fv6->offset;
-    //printf("FILEV6 INODESIZE %d, \n", inode_getsize(&fv6->i_node));
-    //printf("FILEV6 OFFSET %d, \n", fv6->offset);
-    //printf("FILEV6 REM %d, \n", remaining);
-    int j = sector_read(fv6->u->f, inode_findsector(fv6->u, &fv6->i_node, fv6->offset/SECTOR_SIZE), buf);
+    int j = sector_read(fv6->u->f, (uint32_t)inode_findsector(fv6->u, &fv6->i_node, fv6->offset/SECTOR_SIZE), buf);
     if(j == ERR_IO || j == ERR_BAD_PARAMETER){
         return j;
     }
